implemented new word list #49
