+++ config.yaml
@@ -7,9 +7,12 @@ DECK_NAME: "FluentAI"
 # 1 = 100%
 WEIGHTS:
   PHONETIC: 1
-  ORTHOGRAPHIC: 0.33
+  ORTHOGRAPHIC: 0.66
   SEMANTIC: 1
   IMAGEABILITY: 1
+
+# The number of words to consider for each similarity metric
+WORD_LIMIT: 1000
 
 G2P:
   MODEL: "charsiu/g2p_multilingual_byT5_small_100"
